requirements.txt
fastapi
google-genai
google-generativeai
openai
pillow
pydantic
python-multipart
uvicorn